requirements.txt
playwright>=1.41.0
openai>=1.12.0
anthropic>=0.18.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10.0
websockets>=12.0
